feat: 이분검색을 통한 로또번호 비교 메서드 추가 및 Lotto - WinningNumber 상속관계 삭제
