+++ src/main/java/lotto/model/Lotto.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
@@ -8,8 +10,6 @@ public class Lotto {
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
-        checkDuplicate(numbers);
-        isInRange(numbers);
         this.numbers = numbers;
     }
 
@@ -19,23 +19,8 @@ public class Lotto {
         }
     }
 
-    private void checkDuplicate(List<Integer> numbers){
-        HashMap<Integer, Integer> numberOfEach = new HashMap<>();
-        for (int num : numbers){
-            int repeats = numberOfEach.getOrDefault(num, 0);
-            if (repeats >= 1){
-                throw new IllegalArgumentException("중복되는 숫자는 입력할 수 없습니다.");
-            }
-            numberOfEach.put(num, repeats + 1);
-        }
-    }
 
-
-    private void isInRange(List<Integer> numbers){
-        numbers.forEach(num ->{
-            if (num > 45 || num < 1){
-                throw new IllegalArgumentException("숫자는 1과 45사이의 수여야 합니다.");
-            }
-        });
+    public int match(int winnerNumber){
+        return Collections.binarySearch(numbers, winnerNumber);
     }
 }
